+++ src/riverqueue/driver/riversqlalchemy/dbsqlc/pg_misc.sql
@@ -1,2 +1,0 @@
--- name: PGAdvisoryXactLock :exec
-SELECT pg_advisory_xact_lock(@key);