+++ src/riverqueue/driver/riversqlalchemy/dbsqlc/river_job.py
@@ -13,88 +13,19 @@ from . import models
 
 
 JOB_GET_ALL = """-- name: job_get_all \\:many
-SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key
+SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states
 FROM river_job
 """
 
 
 JOB_GET_BY_ID = """-- name: job_get_by_id \\:one
-SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key
+SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states
 FROM river_job
 WHERE id = :p1
 """
 
 
-JOB_GET_BY_KIND_AND_UNIQUE_PROPERTIES = """-- name: job_get_by_kind_and_unique_properties \\:one
-SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key
-FROM river_job
-WHERE kind = :p1
-    AND CASE WHEN :p2\\:\\:boolean THEN args = :p3 ELSE true END
-    AND CASE WHEN :p4\\:\\:boolean THEN tstzrange(:p5\\:\\:timestamptz, :p6\\:\\:timestamptz, '[)') @> created_at ELSE true END
-    AND CASE WHEN :p7\\:\\:boolean THEN queue = :p8 ELSE true END
-    AND CASE WHEN :p9\\:\\:boolean THEN state\\:\\:text = any(:p10\\:\\:text[]) ELSE true END
-"""
-
-
-@dataclasses.dataclass()
-class JobGetByKindAndUniquePropertiesParams:
-    kind: str
-    by_args: bool
-    args: Any
-    by_created_at: bool
-    created_at_begin: datetime.datetime
-    created_at_end: datetime.datetime
-    by_queue: bool
-    queue: str
-    by_state: bool
-    state: List[str]
-
-
-JOB_INSERT_FAST = """-- name: job_insert_fast \\:one
-INSERT INTO river_job(
-    args,
-    created_at,
-    finalized_at,
-    kind,
-    max_attempts,
-    metadata,
-    priority,
-    queue,
-    scheduled_at,
-    state,
-    tags
-) VALUES (
-    :p1\\:\\:jsonb,
-    coalesce(:p2\\:\\:timestamptz, now()),
-    :p3,
-    :p4\\:\\:text,
-    :p5\\:\\:smallint,
-    coalesce(:p6\\:\\:jsonb, '{}'),
-    :p7\\:\\:smallint,
-    :p8\\:\\:text,
-    coalesce(:p9\\:\\:timestamptz, now()),
-    :p10\\:\\:river_job_state,
-    coalesce(:p11\\:\\:varchar(255)[], '{}')
-) RETURNING id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key
-"""
-
-
-@dataclasses.dataclass()
-class JobInsertFastParams:
-    args: Any
-    created_at: Optional[datetime.datetime]
-    finalized_at: Optional[datetime.datetime]
-    kind: str
-    max_attempts: int
-    metadata: Any
-    priority: int
-    queue: str
-    scheduled_at: Optional[datetime.datetime]
-    state: models.RiverJobState
-    tags: List[str]
-
-
-JOB_INSERT_FAST_MANY = """-- name: job_insert_fast_many \\:execrows
+JOB_INSERT_FAST_MANY = """-- name: job_insert_fast_many \\:many
 INSERT INTO river_job(
     args,
     kind,
@@ -104,7 +35,9 @@ INSERT INTO river_job(
     queue,
     scheduled_at,
     state,
-    tags
+    tags,
+    unique_key,
+    unique_states
 ) SELECT
     unnest(:p1\\:\\:jsonb[]),
     unnest(:p2\\:\\:text[]),
@@ -114,10 +47,22 @@ INSERT INTO river_job(
     unnest(:p6\\:\\:text[]),
     unnest(:p7\\:\\:timestamptz[]),
     unnest(:p8\\:\\:river_job_state[]),
+    -- Unnest on a multi-dimensional array will fully flatten the array, so we
+    -- encode the tag list as a comma-separated string and split it in the
+    -- query.
+    string_to_array(unnest(:p9\\:\\:text[]), ','),
 
-    -- Had trouble getting multi-dimensional arrays to play nicely with sqlc,
-    -- but it might be possible. For now, join tags into a single string.
-    string_to_array(unnest(:p9\\:\\:text[]), ',')
+    nullif(unnest(:p10\\:\\:bytea[]), ''),
+    -- Strings of bits are used for the input type here to make sqlalchemy play nicely with bit(8)\\:
+    nullif(unnest(:p11\\:\\:text[]), '')\\:\\:bit(8)
+
+ON CONFLICT (unique_key)
+    WHERE unique_key IS NOT NULL
+      AND unique_states IS NOT NULL
+      AND river_job_state_in_bitmask(unique_states, state)
+    -- Something needs to be updated for a row to be returned on a conflict.
+    DO UPDATE SET kind = EXCLUDED.kind
+RETURNING river_job.id, river_job.args, river_job.attempt, river_job.attempted_at, river_job.attempted_by, river_job.created_at, river_job.errors, river_job.finalized_at, river_job.kind, river_job.max_attempts, river_job.metadata, river_job.priority, river_job.queue, river_job.state, river_job.scheduled_at, river_job.tags, river_job.unique_key, river_job.unique_states, (xmax != 0) AS unique_skipped_as_duplicate
 """
 
 
@@ -132,6 +77,31 @@ class JobInsertFastManyParams:
     scheduled_at: List[datetime.datetime]
     state: List[models.RiverJobState]
     tags: List[str]
+    unique_key: List[memoryview]
+    unique_states: List[str]
+
+
+@dataclasses.dataclass()
+class JobInsertFastManyRow:
+    id: int
+    args: Any
+    attempt: int
+    attempted_at: Optional[datetime.datetime]
+    attempted_by: Optional[List[str]]
+    created_at: datetime.datetime
+    errors: Optional[List[Any]]
+    finalized_at: Optional[datetime.datetime]
+    kind: str
+    max_attempts: int
+    metadata: Any
+    priority: int
+    queue: str
+    state: models.RiverJobState
+    scheduled_at: datetime.datetime
+    tags: List[str]
+    unique_key: Optional[memoryview]
+    unique_states: Optional[Any]
+    unique_skipped_as_duplicate: bool
 
 
 JOB_INSERT_FULL = """-- name: job_insert_full \\:one
@@ -167,7 +137,7 @@ INSERT INTO river_job(
     :p13\\:\\:river_job_state,
     coalesce(:p14\\:\\:varchar(255)[], '{}'),
     :p15
-) RETURNING id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key
+) RETURNING id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states
 """
 
 
@@ -188,79 +158,6 @@ class JobInsertFullParams:
     state: models.RiverJobState
     tags: List[str]
     unique_key: Optional[memoryview]
-
-
-JOB_INSERT_UNIQUE = """-- name: job_insert_unique \\:one
-INSERT INTO river_job(
-    args,
-    created_at,
-    finalized_at,
-    kind,
-    max_attempts,
-    metadata,
-    priority,
-    queue,
-    scheduled_at,
-    state,
-    tags,
-    unique_key
-) VALUES (
-    :p1,
-    coalesce(:p2\\:\\:timestamptz, now()),
-    :p3,
-    :p4,
-    :p5,
-    coalesce(:p6\\:\\:jsonb, '{}'),
-    :p7,
-    :p8,
-    coalesce(:p9\\:\\:timestamptz, now()),
-    :p10,
-    coalesce(:p11\\:\\:varchar(255)[], '{}'),
-    :p12
-)
-ON CONFLICT (kind, unique_key) WHERE unique_key IS NOT NULL
-    -- Something needs to be updated for a row to be returned on a conflict.
-    DO UPDATE SET kind = EXCLUDED.kind
-RETURNING id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, (xmax != 0) AS unique_skipped_as_duplicate
-"""
-
-
-@dataclasses.dataclass()
-class JobInsertUniqueParams:
-    args: Any
-    created_at: Optional[datetime.datetime]
-    finalized_at: Optional[datetime.datetime]
-    kind: str
-    max_attempts: int
-    metadata: Any
-    priority: int
-    queue: str
-    scheduled_at: Optional[datetime.datetime]
-    state: models.RiverJobState
-    tags: List[str]
-    unique_key: Optional[memoryview]
-
-
-@dataclasses.dataclass()
-class JobInsertUniqueRow:
-    id: int
-    args: Any
-    attempt: int
-    attempted_at: Optional[datetime.datetime]
-    attempted_by: Optional[List[str]]
-    created_at: datetime.datetime
-    errors: Optional[List[Any]]
-    finalized_at: Optional[datetime.datetime]
-    kind: str
-    max_attempts: int
-    metadata: Any
-    priority: int
-    queue: str
-    state: models.RiverJobState
-    scheduled_at: datetime.datetime
-    tags: List[str]
-    unique_key: Optional[memoryview]
-    unique_skipped_as_duplicate: bool
 
 
 class Querier:
@@ -288,6 +185,7 @@ class Querier:
                 scheduled_at=row[14],
                 tags=row[15],
                 unique_key=row[16],
+                unique_states=row[17],
             )
 
     def job_get_by_id(self, *, id: int) -> Optional[models.RiverJob]:
@@ -312,80 +210,10 @@ class Querier:
             scheduled_at=row[14],
             tags=row[15],
             unique_key=row[16],
+            unique_states=row[17],
         )
 
-    def job_get_by_kind_and_unique_properties(self, arg: JobGetByKindAndUniquePropertiesParams) -> Optional[models.RiverJob]:
-        row = self._conn.execute(sqlalchemy.text(JOB_GET_BY_KIND_AND_UNIQUE_PROPERTIES), {
-            "p1": arg.kind,
-            "p2": arg.by_args,
-            "p3": arg.args,
-            "p4": arg.by_created_at,
-            "p5": arg.created_at_begin,
-            "p6": arg.created_at_end,
-            "p7": arg.by_queue,
-            "p8": arg.queue,
-            "p9": arg.by_state,
-            "p10": arg.state,
-        }).first()
-        if row is None:
-            return None
-        return models.RiverJob(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-        )
-
-    def job_insert_fast(self, arg: JobInsertFastParams) -> Optional[models.RiverJob]:
-        row = self._conn.execute(sqlalchemy.text(JOB_INSERT_FAST), {
-            "p1": arg.args,
-            "p2": arg.created_at,
-            "p3": arg.finalized_at,
-            "p4": arg.kind,
-            "p5": arg.max_attempts,
-            "p6": arg.metadata,
-            "p7": arg.priority,
-            "p8": arg.queue,
-            "p9": arg.scheduled_at,
-            "p10": arg.state,
-            "p11": arg.tags,
-        }).first()
-        if row is None:
-            return None
-        return models.RiverJob(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-        )
-
-    def job_insert_fast_many(self, arg: JobInsertFastManyParams) -> int:
+    def job_insert_fast_many(self, arg: JobInsertFastManyParams) -> Iterator[JobInsertFastManyRow]:
         result = self._conn.execute(sqlalchemy.text(JOB_INSERT_FAST_MANY), {
             "p1": arg.args,
             "p2": arg.kind,
@@ -396,8 +224,31 @@ class Querier:
             "p7": arg.scheduled_at,
             "p8": arg.state,
             "p9": arg.tags,
+            "p10": arg.unique_key,
+            "p11": arg.unique_states,
         })
-        return result.rowcount
+        for row in result:
+            yield JobInsertFastManyRow(
+                id=row[0],
+                args=row[1],
+                attempt=row[2],
+                attempted_at=row[3],
+                attempted_by=row[4],
+                created_at=row[5],
+                errors=row[6],
+                finalized_at=row[7],
+                kind=row[8],
+                max_attempts=row[9],
+                metadata=row[10],
+                priority=row[11],
+                queue=row[12],
+                state=row[13],
+                scheduled_at=row[14],
+                tags=row[15],
+                unique_key=row[16],
+                unique_states=row[17],
+                unique_skipped_as_duplicate=row[18],
+            )
 
     def job_insert_full(self, arg: JobInsertFullParams) -> Optional[models.RiverJob]:
         row = self._conn.execute(sqlalchemy.text(JOB_INSERT_FULL), {
@@ -437,44 +288,7 @@ class Querier:
             scheduled_at=row[14],
             tags=row[15],
             unique_key=row[16],
-        )
-
-    def job_insert_unique(self, arg: JobInsertUniqueParams) -> Optional[JobInsertUniqueRow]:
-        row = self._conn.execute(sqlalchemy.text(JOB_INSERT_UNIQUE), {
-            "p1": arg.args,
-            "p2": arg.created_at,
-            "p3": arg.finalized_at,
-            "p4": arg.kind,
-            "p5": arg.max_attempts,
-            "p6": arg.metadata,
-            "p7": arg.priority,
-            "p8": arg.queue,
-            "p9": arg.scheduled_at,
-            "p10": arg.state,
-            "p11": arg.tags,
-            "p12": arg.unique_key,
-        }).first()
-        if row is None:
-            return None
-        return JobInsertUniqueRow(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-            unique_skipped_as_duplicate=row[17],
+            unique_states=row[17],
         )
 
 
@@ -503,6 +317,7 @@ class AsyncQuerier:
                 scheduled_at=row[14],
                 tags=row[15],
                 unique_key=row[16],
+                unique_states=row[17],
             )
 
     async def job_get_by_id(self, *, id: int) -> Optional[models.RiverJob]:
@@ -527,81 +342,11 @@ class AsyncQuerier:
             scheduled_at=row[14],
             tags=row[15],
             unique_key=row[16],
+            unique_states=row[17],
         )
 
-    async def job_get_by_kind_and_unique_properties(self, arg: JobGetByKindAndUniquePropertiesParams) -> Optional[models.RiverJob]:
-        row = (await self._conn.execute(sqlalchemy.text(JOB_GET_BY_KIND_AND_UNIQUE_PROPERTIES), {
-            "p1": arg.kind,
-            "p2": arg.by_args,
-            "p3": arg.args,
-            "p4": arg.by_created_at,
-            "p5": arg.created_at_begin,
-            "p6": arg.created_at_end,
-            "p7": arg.by_queue,
-            "p8": arg.queue,
-            "p9": arg.by_state,
-            "p10": arg.state,
-        })).first()
-        if row is None:
-            return None
-        return models.RiverJob(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-        )
-
-    async def job_insert_fast(self, arg: JobInsertFastParams) -> Optional[models.RiverJob]:
-        row = (await self._conn.execute(sqlalchemy.text(JOB_INSERT_FAST), {
-            "p1": arg.args,
-            "p2": arg.created_at,
-            "p3": arg.finalized_at,
-            "p4": arg.kind,
-            "p5": arg.max_attempts,
-            "p6": arg.metadata,
-            "p7": arg.priority,
-            "p8": arg.queue,
-            "p9": arg.scheduled_at,
-            "p10": arg.state,
-            "p11": arg.tags,
-        })).first()
-        if row is None:
-            return None
-        return models.RiverJob(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-        )
-
-    async def job_insert_fast_many(self, arg: JobInsertFastManyParams) -> int:
-        result = await self._conn.execute(sqlalchemy.text(JOB_INSERT_FAST_MANY), {
+    async def job_insert_fast_many(self, arg: JobInsertFastManyParams) -> AsyncIterator[JobInsertFastManyRow]:
+        result = await self._conn.stream(sqlalchemy.text(JOB_INSERT_FAST_MANY), {
             "p1": arg.args,
             "p2": arg.kind,
             "p3": arg.max_attempts,
@@ -611,8 +356,31 @@ class AsyncQuerier:
             "p7": arg.scheduled_at,
             "p8": arg.state,
             "p9": arg.tags,
+            "p10": arg.unique_key,
+            "p11": arg.unique_states,
         })
-        return result.rowcount
+        async for row in result:
+            yield JobInsertFastManyRow(
+                id=row[0],
+                args=row[1],
+                attempt=row[2],
+                attempted_at=row[3],
+                attempted_by=row[4],
+                created_at=row[5],
+                errors=row[6],
+                finalized_at=row[7],
+                kind=row[8],
+                max_attempts=row[9],
+                metadata=row[10],
+                priority=row[11],
+                queue=row[12],
+                state=row[13],
+                scheduled_at=row[14],
+                tags=row[15],
+                unique_key=row[16],
+                unique_states=row[17],
+                unique_skipped_as_duplicate=row[18],
+            )
 
     async def job_insert_full(self, arg: JobInsertFullParams) -> Optional[models.RiverJob]:
         row = (await self._conn.execute(sqlalchemy.text(JOB_INSERT_FULL), {
@@ -652,42 +420,5 @@ class AsyncQuerier:
             scheduled_at=row[14],
             tags=row[15],
             unique_key=row[16],
-        )
-
-    async def job_insert_unique(self, arg: JobInsertUniqueParams) -> Optional[JobInsertUniqueRow]:
-        row = (await self._conn.execute(sqlalchemy.text(JOB_INSERT_UNIQUE), {
-            "p1": arg.args,
-            "p2": arg.created_at,
-            "p3": arg.finalized_at,
-            "p4": arg.kind,
-            "p5": arg.max_attempts,
-            "p6": arg.metadata,
-            "p7": arg.priority,
-            "p8": arg.queue,
-            "p9": arg.scheduled_at,
-            "p10": arg.state,
-            "p11": arg.tags,
-            "p12": arg.unique_key,
-        })).first()
-        if row is None:
-            return None
-        return JobInsertUniqueRow(
-            id=row[0],
-            args=row[1],
-            attempt=row[2],
-            attempted_at=row[3],
-            attempted_by=row[4],
-            created_at=row[5],
-            errors=row[6],
-            finalized_at=row[7],
-            kind=row[8],
-            max_attempts=row[9],
-            metadata=row[10],
-            priority=row[11],
-            queue=row[12],
-            state=row[13],
-            scheduled_at=row[14],
-            tags=row[15],
-            unique_key=row[16],
-            unique_skipped_as_duplicate=row[17],
+            unique_states=row[17],
         )